refactor: FindEmailModal 컴포넌트를 find 폴더로 이동
