+++ src/pages/auth/LoginPage.tsx
@@ -6,7 +6,7 @@ import { PasswordInput } from '@/components/common/PasswordInput'
 import { Button } from '@/components/common/Button'
 import { SocialLoginButton } from '@/components/common/SocialLoginButton'
 import { AlertModal } from '@/components/common/Modal/AlertModal'
-import { FindEmailModal } from '@/components/auth/FindEmailModal'
+import { FindEmailModal } from '@/components/find/FindEmailModal'
 import { useLogin } from '@/features/accounts/login/queries'
 import { meQueries } from '@/features/accounts/me/queries'
 import { useAuthStore } from '@/stores/authStore'
